Sequential parallel thread heap scan with worker threads
Doing parallel thread seq scan on heap table is slower than on single
thread and that is because of need for global lock that needs to be
taken after each fetching of buffer, checking tuple visibility.

To speed up execution, we start parallel worker dedicated for thread
that will fetch buffer and pass them to thread. Thread works with this
page directly and once scan for buffer is done worker will relase it.

HeapTupleSatisfiesVisibility call is also problematic because on some
situations it will try to use SetHintBits on same page and that requires
to have lock on page (which is not true for thread). For this purpose
HeapTupleSatisfiesVisibilityNoHintBits was added which has same logic
but doesn't use SetHintBits.

Preliminatory testing showed that there is small difference between
3,4,.. parallel works so to simplfy logic we use hardcoded rule that
will spawn 1 parallel process if number of blocks in thread is bigger
than 2024 and if bigger 2 parallel workers (threads) will be created.