Change "CREATE EXTENSION duckdb" to "CREATE EXTENSION pg_duckdb"
The README mentioned that the way to install the extension is to use use
`CREATE EXTENSION pg_duckdb` but the actual code didn't actually match
that. Instead the extension was called `duckdb` not `pg_duckdb`.

This aligns both on `pg_duckdb`. I expect some ~~bikeshedding~~ healthy
discussion here. Another solution is obviously to align on `duckdb`
instead.

I definitely want to keep `duckdb` as the schema name, because it's
shorter and the tables/functions stored in this schema will read more
natural. But I think for `CREATE EXTENSION` it doesn't matter much which
of the two we use. The main reason why I went for `pg_duckdb` is because
we call the extension pg_duckdb in our readme all the time, and it
seemed a bit strange to then install something with a different name.