Backport PG17 foreach macros and use them (#89)
For the PG17 release cycle a patch of mine was accepted in upstream
Postgres that adds much easier to use macros for looping over a Postgres
`List`. These new macros completely remove the need to declare a
`ListCell` variable:

<https://github.com/postgres/postgres/commit/14dd0f27d7cd56ffae9ecdbe324965073d01a9ff>

This adds those new macros to our extension so we can use them before
PG17. It also replaces the few loops that we have now.