Cache IsExtensionRegistered (#115)
Doing a btree lookup in `pg_extension` for every query is not free, and
an extension introducing one can (based on my experience) impact
performance for single row lookups significantly.

This introduces a cache which tracks if the extension is registered. I
plan to use this cache in later PRs for more things, such as oids of our
DuckDB functions.

This cache is invalidated whenever an invalidation message is sent for
the "duckdb" schema, which just so happens to be whenever `CREATE/DROP
EXTENSION pg_duckdb` is executed. These invalidation messages are the
mechanism that Postgres uses to let other backends know that some DDL
has occured. We'll probably want to do something else for `ALTER
EXTENSION pg_duckdb UPDATE`, but for now this is good enough (since
there's no upgrades yet). It would have been best if we could receive
invalidation messages for rows in `pg_extension`. But sadly this is not
possible. I submitted a patch to upstream Postgres to hopefully allow
for this in the future (but that will only be in PG18+):

https://www.postgresql.org/message-id/flat/CAGECzQTWm9sex719Hptbq4j56hBGUti7J9OWjeMobQ1ccRok9w%40mail.gmail.com