Add support for CREATE TABLE AS (#130)
This adds support for queries like:

```
CREATE TABLE mytable AS SELECT column00 from read_csv("my_data.csv") AS (column00 int)
```

The table will be then be created in Postgres (not in DuckDB). The main
change that supports this is that we now start "deparsing" the parsed
`Query` type from Postgres to an actual query string, instead of sending
the original string that Postgres received to DuckDB.

This also accidentally fixes some differences such as the DuckDB
`count(*)` aggregate creating a `count_star` vs Postgres its `count`
column. So it reduces the need for #36 a bit. The reason this happens is
because it is deparsed very explicitly like `count(*) AS count`.

For similar reasons this also changes the behaviour of the following
query to only return the `id` and `name` columns, instead of all the
columns. It's arguable whether that's a good thing or not: it's more
verbose if you want those columns but it's also less surprising.

```
SELECT * FROM read_parquet(...) AS (id int, name text);
```

One especially surprising thing with the above query is that
if you used a column name that was not in the parquet file,
then your query would succeed, but you would not actually
get that column back in the result (only the actual columns
in the file).