Speed up running tests when compiled with DUCKDB_BUILD=Debug
Without this change it takes more than a minute to run
`make installcheck` on my machine if I enable DUCKDB_BUILD=Debug.
Because both of these tests take more than thirty seconds. After this
change both execute in less than 2 seconds. Especially the first test
taking so long is confusing, because it seems like the tests are not
working.