Add unmodified ruleutils source files from PG16 and PG17 (#140)
Postgres contains code in the `src/backend/utils/adt/ruleutils.c` file
to build a query string from a `Query *`. We also need to do this to be
able to send a query string to DuckDB, and we started using this
functionality in #130. Sadly this part of the Postgres code is not
extendable at all, and we do need to make some small changes.

So this vendors in the files, one for each supported postgres version.
This commit does not make any changes yet.