Fix using public.read_csv without public in search_path
DuckDB would chokes when Postgres sends it a query with
`public.read_csv` in it. This function is in the `public` schema in
Postgres, but that schema does not even exist in DuckDB.

This starts to fully qualify any of our DuckDB-only functions using the
full DuckDB path e.g. `system.main.read_csv`.