Allow make installcheck to use the -j flag (#139)
When running with `make installcheck -j20`, the regression tests were
started in parallel with the build. This means regression tests will
test the old binaries if any of the source files changed since the last
`make install`. This fixes that by making sure the tests are only
started after the build has completed.