Explicitly convert datum values before filter operation (#194)
Datums can contain a pass-by-value or a pass-by-reference type. We were
simply casting the `Datum` for a pass by value type, but that is not
always correct. Specifically for floating point numbers this would not
always result in the correct floating point number. And int64 values are
pass-by-reference on 32bit systems, but pass-by-value on 64bit systems.
Postgres provides the `DatumGetXyz` macros/functions to work around
these problems. This starts using these functions in
`FilterOperationSwitch` to avoid these casting problems there.