Fix cases where Postgres tuples don't store all columns
Since https://github.com/postgres/postgres/commit/16828d5c0273b4fe5f10f42588005f16b415b2d8 you can add columns with defaults in a way that does not require a table rewrite. The way this is implemented is that the default is stored for the attribute and when reading the tuple, if it has fewer columns than you expect, then you need to fetch the default. We were not doing this, so we would always read those columns as NULL.