Make DuckDB connection a singleton
In #174 we made the DuckDB database a singleton, but a new DuckDB
connection was still being created for every query. This means that any
connection state, such as SET commands or temporary tables don't survive a
single query, which is pretty confusing. This addresses that by also
making the `duckdb::Connection` a singleton, this way a Postgres
connection and a DuckDB connection are linked 1-to-1.