Add duckdb.raw_query function (#203)
For upcoming features, specifically MotherDuck support it is quite
useful to be able to send raw queries to DuckDB to inspect the current
state of the database. Such as `DESCRIBE` or `SELECT * FROM
duckdb_tables();` or even queries to debug the our current `SECRET`
support. This adds a small helper function to do this. The result is
shown in a NOTICE message instead of actually returning them as rows, so
that there's no need for the `AS (some_column int, some_other_column
text)` stuff. We'll probably want a function that returns actual rows
too for use in scripts, but this is a bit easier to use for debugging at
the moment. It's specifically named `duckdb.raw_query`, so we can use
`duckdb.query` for the function that actually returns rows.