Use Postgres its header dependency tracking instead of custom one (#205)
Building the `.depend` file takes ~1.5 seconds on my laptop (in power
saver mode). And this is done every time I execute `make`. Postgres its
Makefile.global has built in support for an equivalent feature that does
not have this downside. The only downside is that it is not enabled on
many Postgres builds, only if you compile with the `--enable-depend`
flag.

This replaces our custom support for dependency tracking with version
copied from the Postgres Makefile. This way even contributors using
Postgres installations that come from their package repositories (which
probably won't have `--enalbe-depend`) won't have to run `make clean`
all the time.