Add PG15 support
This applies changes to pg_ruleutils_15.c that are identical to the ones
that we already made to the PG16 and PG17 versions of these files.

It also notably removes tests involving the psql \bind meta-command,
because those meta-commands are unsupported by PG15 its psql. This is
acceptable because we

Other than that this mainly adds a bunch of #if PG_VERSION_NUM >= 160000
preprocessor statements in places where APIs diverge between versions.