Use -Werror in CI (#237)
The PG15 support that was introduced in #180 introduced a bunch of
warnings when compiling. To avoid getting numb to warnings, this adds
the -Werror to our CI builds, so we catch warnings before merging to
main.

This also starts to ignore the register warning which was introduced by
PG15 support.