Return default value for missing columns (#198)
Postgres can add columns with defaults in a way that does not require a
table
rewrite. The way this is implemented is that the default is stored for
the attribute
and when reading the tuple, if it has fewer columns than you expect,
then
you need to fetch the default. We were not doing this, so we would
always
read those columns as NULL.

Implementation is done in a way to be thread safe. During global state
initialization we populate map structure with non null default values.
When reading tuple from page and in case that tuple has fewer columns
than expected we access this map in thread safe way with
`const_iterator`

---------

Co-authored-by: mkaruza <mkaruza@users.noreply.github.com>