Add DuckDB temporary tables (#231)
This adds support to create temporary tables using the DuckDB in-memory
storage engine. This can be useful for certain ETL workflows and it's an
important stepping-stone to implement MotherDuck support. This is done
by using a new `duckdb` [Table Access Method
(AM)](https://www.postgresql.org/docs/current/tableam.html). However, we
don't really use any of the hooks that the Table AM interface provides
us with. Basically all of these functions are not implement. The reason
for this is that we only really care about the `USING duckdb` syntax as
a way to mark certain tables as being stored in DuckDB. When we notice
such a marker DuckDB completely takes over the execution of the query.

An important implication of this approach is that you cannot create
Postgres indexes on DuckDB tables, which is a feature that the table
access method interface normally gives you for free if you actually
implement the methods. We don't care about this though since DuckDB
wouldn't be able to use these indexes anyway.

Another similar interface that Postgres provides is [FOREIGN
TABLEs](https://www.postgresql.org/docs/current/sql-createforeigntable.html).
This is a way to define a table that is actually backed by an external
data source. This has too many limitation to be useful for what we want
to do though. For example, you cannot use CREATE TABLE AS with a FOREIGN
TABLE. Also, you cannot create temporary foreign tables, which is
exactly the feature that this PR provides.

Note: this PR does not implement CREATE TEMP TABLE AS yet. That will be
done in a follow-up PR, so that this PR does not get too big.

Fixes #224