Improve error when read_xxx functions fail
Previously the error would be shown as a WARNING and then the query
would be run as a fallback using the Postgres executor, which would fail
because the DuckDB `read_csv`/`read_parquet` functions cannot be
executed by the Postgres executor. This is quite confusing as many
people only look at the last ERROR, and not the warnings before.

This is also relevant for MotherDuck support, since the MotherDuck
backed tables can only be read using the DuckDB executor.

Fixes #106