Fix foreach_ptr to work in C++ code (#276)
C++ is more strict in casting void pointers to other pointers than C, so
the upstream `foreach_ptr` implementation doesn't compile on C++ when
it's actually used. I'll make an upstream patch for that shortly, but
for now let's at least fix the problem in our vendored code (all that's
needed is an additional cast).