Add MotherDuck support (#267)
This adds initial MotherDuck support. It supports syncing of table and
schema metadata from MotherDuck to Postgres. Every second a background
worker will check if a MotherDuck catalog has changed to a new version.
If there was a catalog version update, then the changes are synced to
Postgres by dropping and recreating all MotherDuck shell tables related
to that MotherDuck catalog.

The schemas from the default MotherDuck database is merged with schemas
with the same name from Postgres. Schemas from other databases are
encoded in this way:
1. If the db is `awesomedb` schema is `main`, then the postgres schema
name is `ddb$awesomedb`
2. If the db is `awesomedb` schema is `some_other_schema`, then the
postgres schema name is `ddb$awesomedb$some_other_schema`

To enable MotherDuck support all that's needed is adding the following
to your Postgres config:
```
shared_preload_libraries = 'pg_duckdb'
duckdb.background_worker_db = '<the db in which you do CREATE EXTENSION>'
duckdb.motherduck_token = '<your token>`
```

This also fixes a problem with temporary duckdb tables that were not
explicitely dropped before the backend was terminated. The oids of those
tables would stay in the `duckdb.tables` table indefinitely. To fix that
this now tracks temporary table oids in backend-local in-memory set.

TODO:
- [x] Check MotherDuck catalog version to see if Postgres catalog needs
to be updated, and do nothing if version is the same.
- [ ] ~~Only update tables that have changed since the last sync
(optional?)~~ (creating separate issue for 0.2.0 for this)
- [ ] ~~Create the tables as the duckdb.motherduck_postgres_user~~ (I'm
going to make a separate PR for this)
- [x] Change to use eager committing of transactions, so Postgres locks
are not held for long periods of time.
- [x] Disallow GRANT on table
- [x] Disallow creation of `ddb$` schemas
- [x] Check that no `ddb$` schemas exist in extension creation script
- [x] Automatically drop all `ddb$` schemas when dropping the extension
- [x] Make the background worker handle create+drop extension sensibly