Pre-configure docker image to use pg_duckdb (#305)
This slightly modifies the default Postgres docker image so that when it
initializes the database it will automatically configure `pg_duckdb`. It
also uses the `postgres` user to start the container, so that `docker
exec ... psql` just works without having to specify `-U postgres`

This also uses a more selective COPY. I often have some random stuff in
my top repo directory, such as gigabytes of TPCDS csv files. This makes
sure those are not copied into the container.