Rename duckdb.execution to duckdb.force_execution
Many people won't need to turn this setting on anymore now that we automatically upgrade to DuckDB execution in many cases. So this changes the name of `duckdb.execution` to `duckdb.force_execution` to make that obvious. It also removes the WARNING during installation and updates the relevant section of the README.