Optimize heap reading and filtering
Currently we use map to match duckdb output column idx to postgres column
id. For column filter we would also do lookup for each column. This is
not needed as this information are not changed during scan of heap
tables.
Optimization is based on idea to construct vector that holds these mapping
information so map lookups are not needed anymmore. For filtering, same
logic is applied, now we would have vector of column filters.