Lookup is_null correctly in case of pushed down filters
While running TPCDS queries a few of them crashed Postgres when using
`duckdb.force_execution = true`. It turns out that happened because it
would look at the NULL value for the wrong field. This fixes that as
well as including a minimal reproducing test of the issue. It took me
quite a while to figure out what was going on because the lookup code is
pretty complex. I'll open another issue to refactor/comment it a bit
because I have some ideas on how to improve it.