Support non-superuser usage of pg_duckdb
DuckDB is not really designed for use in a multi-user system, and Postgres is a multi-user system. This is visible through at least two ways:

1. An important security concern here is that secrets are usable by all users, and the same concern applies to MotherDuck server-side secrets. We can possibly work around these issues.
2. DuckDB doesn't enforce any permissions when accessing tables, but Postgres applies permissions on tables.

So security wise it's not straight forward to allow non-superusers in Postgres to use DuckDB execution, without allowing them to escalate privileges somehow. Allowing regular users to run arbitrary SQL in DuckDB is a huge security no-go, since they can bypass any permission checks from any table in Postgres.

This concern is addressed in five ways:
1. By default only allow superuser to run DuckDB queries. Security wise this would be enough on its own, but since this makes usage on production systems near-impossible we add some more options.
2. Allow configuring a `duckdb.postgres_role` setting which contains a role name that can also use DuckDB execution. A role in Postgres can be used as a group and can be assigned to multiple users.
3. Have Postgres check permissions on tables used in a query before forwarding them DuckDB.
4. Don't support queries where Row Level Security (RLS) needs to be enforced when reading tables.
5. Don't allow non-superusers to run any functions in the `duckdb`
   schema that trigger manually constructed queries. So no `duckdb.raw_query()`, and no `duckdb.install_extension()`. I haven't deeply investigated these for possible SQL injection yet, but it's `raw_query` is SQL injection by design, and `LoadExtensions` is currently not escaping the extension name yet before passing it to queries. So for now these are super user only to decrease attack surface.

NOTE: This means we have to be extra careful to avoid SQL injection when constructing DuckDB queries, because that would allow a non-superuser with the `duckdb.postgres_role` to execute arbitrary SQL in DuckDB, and thus read arbitrary Postgres tables. Luckily even if this happens currently, it's limited to reading arbitrary data, not modifying, because we don't support writes to Postgres tables from DuckDB yet.