Don't get page if buffer was released
There is edge case when we relase buffer and also have complete output
vector filled. In that case we would call `BufferGetPage` with
InvalidBuffer as parameter. Current check is only to check block which
can be valid. We should check do we need to read next buffer or not
instead.