Create MotherDuck tables as duckdb.postgres_user
Before we were always creating tables as the bootstrap superuser, this means that the `duckdb.postgres_user` was not able to access them. This changes that and also protects against a type of attack where someone with access to the MotherDuck account, but not full access to the Postgres server can create tables in any Postgres schema by creating them in MotherDuck and waiting for them to be synced. Now by setting `duckdb.postgres_user` only schemas that the `duckdb.postgres_user` can access will allow tables to be created in it by the background sync.