Resolve security issues reported by pgspot (#343)
It's really easy to make critical security mistakes in the SQL scripts
of an extension. Luckily [pgspot][1] can help finding those. There were
a few false positives because the tool did not realize that the
search_path `duckdb, pg_catalog, pg_temp` was also safe. I'll probably
make a PR to pgspot to allow specifying additional safe schemas. For now
I just followed its advice to get to 0 errors, and I think the script is
still pretty readable.

I also changed the script to use `@extschema@` whenever we were
previously storing a function in the `public` schema. This way we can
more easily choose to make the extension relocatable in the future. Or
at least allow people to store the these objects in a schema of their
choice by doing `CREATE EXTENSION pg_duckdb WITH SCHEMA
some_other_schema`. For now I forced the schema to `public` in the
control file. Because we don't have tests for this multi schema stuff
and we might not think its worth the effort in the long run.

[1]: https://github.com/timescale/pgspot

Fixes #250

---------

Co-authored-by: Jonathan Dance (JD) <wuputah@users.noreply.github.com>