Correctly set missing_ok to false in the GRANT trigger (#344)
We expect this RangeVarGetRelid call to actually find the relid, since
the event trigger only fires when the grant was successfully done by
Postgres. But we were passing it `missing_ok = false` as the last
argument. In practice this didn't matter, because it always found it.
But in an earlier version of #343 I broke that, which would then cause a
crash later on. This passing `missing_ok = true` instead so that next
someone breaks something at least it will fail with a nice error instead
of a crash.