Fix glibc assertion
We were using `reserve` instead of `resize` to size our vector `values`
and `nulls` vectors. This meant that we were reading out of bounds of
the vector its size, but still within its capacity.

This also enables glibc assertions whenever they are available in Debug
builds because that is how this issue was caused.

Fixes #398