Cache DuckDB connection object and hook into PG transaction life cycle (#413)
Instead of creating a new `duckdb::Connection` over and over again, this
starts caching the DuckDB connection. For this to work correctly in case
of errors we need to link DuckDB and Postgres its transaction life cycle
together. Otherwise DuckDB state from a previous PG query might
accidentally survive until the next PG. This is also in preparation of
supporting multi-statement transactions.

There was one more problem to solve: This PR changes the lifecycle of
the `duckdb::Transaction`.
Before the lifecycle of the different components was as follows, from
long-lived to short-lived:
1. Postgres transaction
2. Postgres query
3. DuckDB transaction
4. DuckDB query

Now 2 and 3 are swapped:
1. Postgres transaction
2. DuckDB transaction
3. Postgres query
4. DuckDB query

This causes problems, because the `Relation` from the Postgres relcache
should always be closed before the end of the postgres query. Previously
we bound `SchemaItems` cache to the lifecycle of the DuckDB transaction,
which now outlives the postgres query. To solve this we change its
lifecycle to be bound the the DuckDB query.