Add PG14 support (#397)
PG14 support needs fairly minimal changes. Notable changes are:

`\getenv` does not exist in psql for PG14, so we replace it's usages
with `\set` commands.

`ALTER TABLE ... SET ACCESS METHOD` does not exist in PG14, so we move
the `ALTER TABLE` tests to `pycheck`.

The PG14 ruleutils file misses some functions that exist in more recent
versions. So we backport `GET_PRETTY_FLAGS` (which we changed) and
`pg_get_querydef` (which we used).