Support multi-statement transactions (#433)
This starts supporting multi-statement transactions. It imposes one
important restriction that is necessary to ensure the expected ACID
guarantees: You cannot write to both a Postgres table and a DuckDB table
in the same transaction.

This also means that we do not allow doing DuckDB DDL inside a
transaction, because that needs to write both to DuckDB and to the
Postgres catalogs.

Either of these restrictions might be removed in a follow up PR.
Especifically the DDL restriction is one that we would like to remove
(and probably can). But for now this PR only implements the obviously
correct case, and any further removal of restrictions will need some
design/tradeoffs discussion.

Fixes #419