Correctly populate target_list of CustomScan
`EXPLAIN VERBOSE` was erroring out when the plan contained one of our
`CustomScan` nodes. The reason for this was that we were not populating
the `targetlist` and `custom_scan_tlist` fields in a way that PG
understood so it was recursing infinitely into the same function. This
changes how we populate those fields to match what PG expects.

Fixes #465
Fixes #466