Correctly populate targetlist of CustomScan (#468)
`EXPLAIN VERBOSE` was erroring out when the plan contained one of our
`CustomScan` nodes. The reason for this was that we were not populating
the `targetlist` and `custom_scan_tlist` fields in a way that PG
understood so it was recursing infinitely into the same function. This
changes how we populate those fields to match what PG expects.

Sadly the additional info in the plan is currently not super useful, and
arguably wrong. It will always show the following line in the explain
plan.
```
   Output: duckdb_scan.explain_key, duckdb_scan.explain_value
```

To fix that we shouldn't include the EXPLAIN part when sending it to
DuckDB for initial planning. Sadly that then breaks the code that we use
for executing the query. So for now we're happy with not throwing an
error, and that we're populating `targetlist` correctly (which might
have some more benefits that we didn't know of).

Fixes #465
Fixes #466
Fixes #463