INSERT ... SELECT or multi-row VALUES errors when using default values (#448)
If it's an INSERT ... SELECT or multi-VALUES INSERT, then
`rewriteTargetListIU` changes the query to put the `DEFAULT` expression
into the targetlist of INSERT command. Since `pg_get_querydef` normally
doesn't have to deal with rewritten queries, it generates a wrong query
because of this. So we need to check if the expression is coming from
the `DEFAULT` clause or the actual query. There's a bunch of different
cases that this code needs to handle which are all covered in the tests.

Fixes #445