Correct handling parameter (#491)
DuckDB complains when it receives parameters that were not actually used
in the query. From Postgres we get all supplied parameters though. So
this changes the code to filter the unused ones out. To do this we lot
DuckDB parse the query and tell us which parameters it expects, and then
we remove the unexpected ones.

Fixes #411 
Fixes #234

Co-authored-by: Zhou Sun <zhou@mooncakelabs.com>