Bump clang-format and ruff versions to latest (#496)
In #98 we started using the same version of clang-format that DuckDB
uses. That version is pretty ancient though. So it misses a bunch of
features that we might want to use, and also has a bunch of bugs which
result in sub-optimal formatting. This updates `clang-format` to the
latest version, and while we're at it we also update `ruff`.

To be able to reformat all existing files, this also adds a `format-all`
target to our `Makefile`.

The problem was called out by @dpxcc in #493, but I've been also running
into this myself for a while.