Use new clang-format in CI (#518)
In #496 we started using the newest clang-format to format our files,
but CI was still installing the old version. This meant that we were not
catching some unformatted files correctly. An example of this being:
https://github.com/duckdb/pg_duckdb/pull/511#discussion_r1901376718

This starts using the correct clang-format version in CI too and formats
any incorrectly formatted files.