Create dedicated cache FS (#550)
Today, non super-users are going to face an error when querying S3
objects like:

```
ERROR:  (PGDuckDB/CreatePlan) Prepared query returned an error: 'Permission Error: File system LocalFileSystem has been disabled by configuration
```

The reason is that the current cache mechanism uses the DB's
`FileSystem` which non-super users cannot access.

To fix it, we create a dedicated FS that the extension is allowed to
access (but not the user directly).