Use duckdb functions without having to specify the full type signature (#531)
Currently when you use `read_parquet`/`read_csv`/etc they require
specifying the full type signature of the function call. This is quite
cumbersome, especially when using `SELECT * FROM read_parquet(...) AS
(...)`. Using `SELECT *` its main advantage is that you don't have to
specify all the columns.

This changes that by instead allowing the following new syntax:

```sql
SELECT * FROM read_parquet('people.parquet') r WHERE r['age'] > 21;
```

This syntax is made possible by using two clever hacks to solve two
different problems:

The first problem is that Postgres its parser needs to consider the
query valid, which means it has to resolve all the types. Sadly, we
cannot hook into the type resolution as an extension. Instead the types
need to be resolved fully from information in the Postgres catalogs. To
do so we create a new type for our extension: `duckdb.unresolved_type`.
This type only exists for the Postgres parser and should never be used
explicitly. We add catalog entries that allow this type be casted
explicitly to any type that's supported by pg_duckdb. Using explicit
casts always is quite annoying though. So we also define various
operators and functions for this type, so that things like `r['age'] >
21` are allowed.

The second problem is that the Postgres parser replaces the `*` in
`SELECT *` with an expanded column list. So the query above will look as
follows once we receive it (`*` replaced with `r`):

```sql
SELECT r FROM read_parquet('people.parquet') r WHERE r['age'] > 21;
```

To resolve this problem we let `read_parquet` return a `duckdb.row`,
which we then replace with a `*` again before in our
`pgduckdb_get_querydef` function. This might sound pretty simple but
there's tricks that need to be done for subqueries that return both a
`duckdb.row` type and some other columns.